requirements.txt
# Core LangChain
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# LLM providers (pick one or both)
langchain-huggingface>=0.1.0
transformers>=4.40.0
accelerate>=0.30.0

# Embeddings
sentence-transformers>=2.7.0

# Vector database
faiss-cpu>=1.8.0
# chromadb>=0.5.0   # optional alternative

# Document loading
pypdf>=4.2.0
unstructured>=0.14.0
markdown>=3.6

# Environment variables
python-dotenv>=1.0.1

# Utilities
tqdm>=4.66.0
numpy>=1.26.0
PyPDFLoader
chromadb
langchain-chroma>=0.1.0